Split huge test class into multiple smaller classes
